requirements.txt
anyio==3.5.0
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.76.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
mysql-connector-python==8.0.29
protobuf==3.20.1
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.18.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
zipp==3.8.0
